# docs/swagger.json
{
  "components": {
    "schemas": {
      "data": {
        "title": "data Schema",
        "properties": { "data": { "$ref": "#/components/schemas/Wallet" } },
          "title": "properties Schema",
        "type": "object"
      },
      "ErrorResponse": {
        "title": "ErrorResponse Schema",
        "description": "Application error response",
        "properties": {
          "code": {
            "enum": [400, 401, 404, 500, 502, 422, 403, 409, 429, 501],
            "example": 400,
            "type": "integer"
          },
          "error": { "example": "field: required", "type": "string" },
          "message": {
            "enum": [
              "Invalid request parameters",
              "Authorization failed",
              "Resource not found",
              "Internal server error",
              "Database error occurred",
              "External service error",
              "Error rendering response",
              "Access forbidden",
              "Resource conflict",
              "Too many requests",
              "Unsupported operation"
            ],
            "example": "Invalid request parameters",
            "type": "string"
          },
          "type": { "$ref": "#/components/schemas/ErrorType" }
        },
        "type": "object"
      },
      "ErrorType": {
        "title": "ErrorType Schema",
        "type": "string",
        "x-enum-varnames": [
          "ErrorTypeValidation",
          "ErrorTypeDatabase",
          "ErrorTypeAuthorization",
          "ErrorTypeNotFound",
          "ErrorTypeInternal",
          "ErrorTypeExternalService",
          "ErrorTypeRender",
          "ErrorTypeForbidden",
          "ErrorTypeConflict",
          "ErrorTypeRateLimit",
          "ErrorTypeUnsupported"
        ]
      },
      "Response": {
        "title": "Response Schema",
        "allOf": [{ "$ref": "#/components/schemas/data" }],
        "description": "Standard API response wrapper",
        "properties": {
          "data": {},
          "message": {
            "enum": [
              "Success",
              "Resource created successfully",
              "Resource updated successfully",
              "Resource deleted successfully"
            ],
            "example": "Success",
            "type": "string"
          },
          "meta": {
            "properties": {
              "count": { "type": "integer" },
              "limit": { "type": "integer" },
              "next_token": { "type": "string" },
              "query": { "type": "string" }
            },
            "type": "object"
          },
          "status": {
            "enum": [200, 202, 204],
            "example": 200,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ContactUpdatePayload": {
        "title": "ContactUpdatePayload Schema",
        "description": "Payload for updating an existing contact",
        "properties": {
          "addressLine1": {
            "example": "123 Main St",
            "maxLength": 255,
            "type": "string"
          },
          "addressLine2": {
            "example": "Suite 100",
            "maxLength": 255,
            "type": "string"
          },
          "city": { "example": "New York", "maxLength": 255, "type": "string" },
          "country": {
            "example": "US",
            "format": "iso-3166-1-alpha-2",
            "type": "string"
          },
          "email": {
            "example": "john.doe@example.com",
            "format": "email",
            "type": "string"
          },
          "name": {
            "example": "John Doe",
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "phone": {
            "example": "+1-555-123-4567",
            "format": "phone",
            "maxLength": 20,
            "type": "string"
          },
          "stateProvince": {
            "example": "NY",
            "maxLength": 255,
            "type": "string"
          },
          "tags": {
            "example": [
              "123e4567-e89b-12d3-a456-426614174000",
              "123e4567-e89b-12d3-a456-426614174001"
            ],
            "items": { "type": "string" },
            "type": "array",
            "uniqueItems": false
          },
          "zipPostalCode": {
            "example": "10001",
            "format": "zip-code",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Wallet": {
        "title": "Wallet Schema",
        "description": "A wallet entity",
        "properties": {
          "balance": { "example": 100.5, "type": "number" },
          "createdAt": { "example": "2023-01-01T00:00:00Z", "type": "string" },
          "currency": { "example": "USD", "type": "string" },
          "name": { "example": "My Wallet", "type": "string" },
          "projectId": {
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "type": "string"
          },
          "tags": {
            "items": { "type": "string" },
            "type": "array",
            "uniqueItems": false
          },
          "updatedAt": { "example": "2023-01-01T00:00:00Z", "type": "string" },
          "userId": {
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "type": "string"
          },
          "walletId": {
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Bearer token authentication",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": { "email": "support@example.com", "name": "API Support" },
    "description": "REST API for expense tracking application with user management",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "Expense Tracker API",
    "version": "1.0"
  },
...
  "paths": {
...
    "/contacts/{id}": {
      "delete": {
        "description": "Deletes a Contact by ID",
        "operationId": "DeleteContact",
        "parameters": [
          {
            "description": "Contact ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Response" }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Delete a Contact",
        "tags": ["Contacts"]
      },
      "get": {
        "description": "Retrieves a Contact by ID",
        "operationId": "GetContact",
        "parameters": [
          {
            "description": "Contact ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Get a Contact",
        "tags": ["Contacts"]
      },
      "put": {
        "description": "Updates an existing Contact",
        "operationId": "UpdateContact",
        "parameters": [
          {
            "description": "Contact ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContactUpdatePayload" }
            }
          },
          "description": "Contact update request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Update a Contact",
        "tags": ["Contacts"]
      }
    },
...
    "/users/me": {
      "get": {
        "description": "Retrieves the profile of the currently authenticated user",
        "operationId": "GetUser",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Get authenticated user profile",
        "tags": ["Users"]
      }
    },
...
  },
  "openapi": "3.1.0",
  "servers": [{ "url": "localhost:8080/api/v1" }]
}